+++ examples/custom-renderer-codelab/src/index.js
@@ -33,7 +33,7 @@ const runCode = () => {
   const code = javascriptGenerator.workspaceToCode(ws);
   codeDiv.innerText = code;
 
-  outputDiv.innerHTML = '';
+  outputDiv.textContent = '';
 
   eval(code);
 };