+++ plugins/field-colour/test/golden/golden.lua
@@ -1,0 +1,35 @@
+function colour_rgb(r, g, b)
+  r = math.floor(math.min(100, math.max(0, r)) * 2.55 + .5)
+  g = math.floor(math.min(100, math.max(0, g)) * 2.55 + .5)
+  b = math.floor(math.min(100, math.max(0, b)) * 2.55 + .5)
+  return string.format("#%02x%02x%02x", r, g, b)
+end
+
+function colour_blend(colour1, colour2, ratio)
+  local r1 = tonumber(string.sub(colour1, 2, 3), 16)
+  local r2 = tonumber(string.sub(colour2, 2, 3), 16)
+  local g1 = tonumber(string.sub(colour1, 4, 5), 16)
+  local g2 = tonumber(string.sub(colour2, 4, 5), 16)
+  local b1 = tonumber(string.sub(colour1, 6, 7), 16)
+  local b2 = tonumber(string.sub(colour2, 6, 7), 16)
+  local ratio = math.min(1, math.max(0, ratio))
+  local r = math.floor(r1 * (1 - ratio) + r2 * ratio + .5)
+  local g = math.floor(g1 * (1 - ratio) + g2 * ratio + .5)
+  local b = math.floor(b1 * (1 - ratio) + b2 * ratio + .5)
+  return string.format("#%02x%02x%02x", r, g, b)
+end
+
+
+local _ = '#ff0000'
+
+local _ = '#3333ff'
+
+local _ = string.format("#%06x", math.random(0, 2^24 - 1))
+
+local _ = colour_rgb(100, 50, 0)
+
+local _ = colour_rgb(0, 1, 20)
+
+local _ = colour_blend('#ff0000', '#3333ff', 0.5)
+
+local _ = colour_blend('#000000', '#ffffff', 0.3)