+++ plugins/migration/bin/rename.js
@@ -192,7 +192,9 @@ class VersionRenamer {
       //   moduleA.exportA -> moduleC.exportB
       // And we performed the module rename first, we wouldn't be able to detect
       // the export rename.
-      this.renamings_.push({old: oldModulePath, new: newModulePath});
+      if (newModulePath !== oldModulePath) {
+        this.renamings_.push({old: oldModulePath, new: newModulePath});
+      }
     }
   }