+++ plugins/block-plus-minus/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.0.6](https://github.com/google/blockly-samples/compare/@blockly/block-plus-minus@7.0.5...@blockly/block-plus-minus@7.0.6) (2024-05-09)
+
+
+### Bug Fixes
+
+* Fix incorrect Blockly imports ([#2339](https://github.com/google/blockly-samples/issues/2339)) ([9af40b9](https://github.com/google/blockly-samples/commit/9af40b9ca075275af2b48cedcc1750d458084eb3))
+
+
+
+## [7.0.5](https://github.com/google/blockly-samples/compare/@blockly/block-plus-minus@7.0.4...@blockly/block-plus-minus@7.0.5) (2024-04-25)
+
+**Note:** Version bump only for package @blockly/block-plus-minus
+
+
+
+
+
+## [7.0.4](https://github.com/google/blockly-samples/compare/@blockly/block-plus-minus@7.0.3...@blockly/block-plus-minus@7.0.4) (2024-04-04)
+
+**Note:** Version bump only for package @blockly/block-plus-minus
+
+
+
+
+
+## [7.0.3](https://github.com/google/blockly-samples/compare/@blockly/block-plus-minus@7.0.2...@blockly/block-plus-minus@7.0.3) (2024-03-28)
+
+**Note:** Version bump only for package @blockly/block-plus-minus
+
+
+
+
+
 ## [7.0.2](https://github.com/google/blockly-samples/compare/@blockly/block-plus-minus@7.0.1...@blockly/block-plus-minus@7.0.2) (2024-02-08)
 
 **Note:** Version bump only for package @blockly/block-plus-minus