+++ plugins/disable-top-blocks/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.14](https://github.com/google/blockly-samples/compare/@blockly/disable-top-blocks@0.4.13...@blockly/disable-top-blocks@0.4.14) (2024-05-09)
+
+**Note:** Version bump only for package @blockly/disable-top-blocks
+
+
+
+
+
+## [0.4.13](https://github.com/google/blockly-samples/compare/@blockly/disable-top-blocks@0.4.12...@blockly/disable-top-blocks@0.4.13) (2024-04-25)
+
+**Note:** Version bump only for package @blockly/disable-top-blocks
+
+
+
+
+
+## [0.4.12](https://github.com/google/blockly-samples/compare/@blockly/disable-top-blocks@0.4.11...@blockly/disable-top-blocks@0.4.12) (2024-04-04)
+
+**Note:** Version bump only for package @blockly/disable-top-blocks
+
+
+
+
+
+## [0.4.11](https://github.com/google/blockly-samples/compare/@blockly/disable-top-blocks@0.4.10...@blockly/disable-top-blocks@0.4.11) (2024-03-28)
+
+**Note:** Version bump only for package @blockly/disable-top-blocks
+
+
+
+
+
 ## [0.4.10](https://github.com/google/blockly-samples/compare/@blockly/disable-top-blocks@0.4.9...@blockly/disable-top-blocks@0.4.10) (2024-02-08)
 
 **Note:** Version bump only for package @blockly/disable-top-blocks