+++ plugins/field-date/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.0.16](https://github.com/google/blockly-samples/compare/@blockly/field-date@8.0.15...@blockly/field-date@8.0.16) (2024-05-09)
+
+**Note:** Version bump only for package @blockly/field-date
+
+
+
+
+
+## [8.0.15](https://github.com/google/blockly-samples/compare/@blockly/field-date@8.0.14...@blockly/field-date@8.0.15) (2024-04-25)
+
+**Note:** Version bump only for package @blockly/field-date
+
+
+
+
+
+## [8.0.14](https://github.com/google/blockly-samples/compare/@blockly/field-date@8.0.13...@blockly/field-date@8.0.14) (2024-04-04)
+
+**Note:** Version bump only for package @blockly/field-date
+
+
+
+
+
+## [8.0.13](https://github.com/google/blockly-samples/compare/@blockly/field-date@8.0.12...@blockly/field-date@8.0.13) (2024-03-28)
+
+**Note:** Version bump only for package @blockly/field-date
+
+
+
+
+
 ## [8.0.12](https://github.com/google/blockly-samples/compare/@blockly/field-date@8.0.11...@blockly/field-date@8.0.12) (2024-02-16)
 
 **Note:** Version bump only for package @blockly/field-date